+++ client/src/test/java/replicant/messages/ChangeSetTest.java
@@ -63,12 +63,12 @@ public class ChangeSetTest
     };
     final EntityChange[] entityChanges = new EntityChange[]{
       EntityChange.create( 1, 1, new String[ 0 ] ),
-      EntityChange.create( 2, 1, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 3, 1, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 4, 1, new String[ 0 ] ),
-      EntityChange.create( 33, 2, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 34, 3, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 1, 4, new String[ 0 ] )
+      EntityChange.create( 1, 2, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 1, 3, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 1, 4, new String[ 0 ] ),
+      EntityChange.create( 2, 33, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 3, 34, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 4, 1, new String[ 0 ] )
     };
 
     final ChangeSet changeSet =
@@ -121,11 +121,11 @@ public class ChangeSetTest
   {
     final EntityChange[] entityChanges = new EntityChange[]{
       EntityChange.create( 1, 1, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 2, 1, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 3, 1, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 4, 1, new String[ 0 ] ),
-      EntityChange.create( 33, 2, new String[ 0 ], new EntityChangeDataImpl() ),
-      EntityChange.create( 34, 3, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 1, 2, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 1, 3, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 1, 4, new String[ 0 ] ),
+      EntityChange.create( 2, 33, new String[ 0 ], new EntityChangeDataImpl() ),
+      EntityChange.create( 3, 34, new String[ 0 ], new EntityChangeDataImpl() ),
       EntityChange.create( 1, 1, new String[ 0 ] )
     };
 
@@ -135,7 +135,7 @@ public class ChangeSetTest
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class, changeSet::validate );
     assertEquals( exception.getMessage(),
-                  "Replicant-0014: ChangeSet 77 contains multiple EntityChange messages for the entity of type 1 and id 1." );
+                  "Replicant-0014: ChangeSet 77 contains multiple EntityChange messages with the id '1.1'." );
   }
 
   @Test